+++ example/src/screens/Settings/index.tsx
@@ -1,31 +1,50 @@
-import { useMemo, useRef, useState } from 'react';
+import { useEffect, useMemo, useRef, useState } from 'react';
 import {
   Button,
   ImageBackground,
   Modal,
   ScrollView,
-  StyleSheet,
   Text,
   TouchableOpacity,
+  useColorScheme,
   View,
 } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import { BlurTarget, BlurView } from '@danielsaraldi/react-native-blur-view';
+import {
+  BlurTarget,
+  BlurView,
+  VibrancyView,
+} from '@danielsaraldi/react-native-blur-view';
+import Animated, {
+  useAnimatedProps,
+  useSharedValue,
+  withRepeat,
+  withTiming,
+} from 'react-native-reanimated';
 import { useBlur } from '../../hooks';
 import { PORSCHE_ARCHITECTURE } from '../../assets';
-import { BLUR_RADIUS_DATA } from '../../constants';
+import { BLUR_RADIUS_DATA, BLUR_UI_MODES } from '../../constants';
 import { makeStyles } from './styles';
 import { isIos } from '../../utils';
+
+const AnimatedBlurView = Animated.createAnimatedComponent(BlurView);
+const AnimatedVibrancyView = Animated.createAnimatedComponent(VibrancyView);
 
 export function Settings() {
   const [isOpenModal, setIsOpenModal] = useState<boolean>(false);
 
   const targetRef = useRef<View | null>(null);
   const scrollTargetRef = useRef<View | null>(null);
+  const colorScheme = useColorScheme();
   const { top, bottom } = useSafeAreaInsets();
   const { blurType, effectStyle, radius, isDark, onRadius } = useBlur();
 
-  const color = blurType.includes('dark') ? 'white' : 'black';
+  const radiusAnimation = useSharedValue(0.01);
+
+  const colorByType = blurType.includes('dark') ? 'white' : 'black';
+  const isDarkMode = colorScheme === 'dark';
+  const isUIMode = BLUR_UI_MODES.some((mode) => blurType === mode);
+  const color = isUIMode ? (isDarkMode ? 'white' : 'black') : colorByType;
   const defaultMessage = isIos
     ? 'Default blur type is light, radius is 10 and vibrancy effect style is label.'
     : 'Default blur type is light and default radius is 10.';
@@ -47,7 +66,6 @@ export function Settings() {
               radius={blurRadius}
               type={blurType}
               style={styles.centralize}
-              reducedTransparencyFallbackColor="#F1F1F1"
             >
               <Text style={[styles.itemText, { color }]}>{label}</Text>
             </BlurView>
@@ -57,15 +75,23 @@ export function Settings() {
     [blurType, styles, color, onRadius]
   );
 
+  const animatedProps = useAnimatedProps(() => ({
+    radius: radiusAnimation.get() * 100,
+  }));
+
+  useEffect(() => {
+    radiusAnimation.set(() =>
+      withRepeat(withTiming(1, { duration: 2000 }), -1, true)
+    );
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
-    <BlurTarget
-      ref={targetRef}
-      style={[styles.expand, StyleSheet.absoluteFill]}
-    >
+    <BlurTarget ref={targetRef} style={[styles.expand, styles.absoluteFill]}>
       <View style={styles.expand}>
-        <BlurTarget ref={scrollTargetRef} style={StyleSheet.absoluteFill}>
+        <BlurTarget ref={scrollTargetRef} style={styles.absoluteFill}>
           <ImageBackground
-            style={StyleSheet.absoluteFill}
+            style={styles.absoluteFill}
             source={PORSCHE_ARCHITECTURE}
             resizeMode="cover"
           />
@@ -85,18 +111,13 @@ export function Settings() {
               blurTarget={scrollTargetRef}
               type={blurType}
               radius={radius}
-              style={StyleSheet.absoluteFill}
+              style={styles.absoluteFill}
             />
 
-            <Text
-              style={[
-                styles.configurationText,
-                isDark && styles.configurationTextDark,
-              ]}
-            >
+            <Text style={[styles.configurationText, { color }]}>
               Explore radius and type configurations to customize the blur
-              effect. Adjust the settings to see how they impact the appearance
-              of the blur on both Android and iOS platforms.
+              effect ✨{'\n'}Adjust the settings to see how they impact the
+              appearance of the blur on both Android and iOS platforms 🌫️
             </Text>
           </View>
 
@@ -106,7 +127,34 @@ export function Settings() {
 
           {renderBlurRadius}
 
-          <View style={[styles.header, styles.modalMarginTop]}>
+          <View style={[styles.header, styles.marginTop]}>
+            <Text style={styles.headerText}>Blur + Reanimated</Text>
+          </View>
+
+          <View style={styles.item}>
+            <AnimatedBlurView
+              blurTarget={scrollTargetRef}
+              type={blurType}
+              style={styles.centralize}
+              animatedProps={animatedProps}
+            >
+              <Text style={[styles.itemText, { color }]}>Blur Animation</Text>
+            </AnimatedBlurView>
+          </View>
+
+          {isIos && (
+            <View style={styles.item}>
+              <AnimatedVibrancyView
+                type={blurType}
+                style={styles.centralize}
+                animatedProps={animatedProps}
+              >
+                <Text style={styles.itemText}>Vibrancy Animation</Text>
+              </AnimatedVibrancyView>
+            </View>
+          )}
+
+          <View style={[styles.header, styles.marginTop]}>
             <Text style={styles.headerText}>Modal</Text>
           </View>
 
@@ -120,7 +168,6 @@ export function Settings() {
               radius={radius}
               type={blurType}
               style={styles.centralize}
-              reducedTransparencyFallbackColor="#F1F1F1"
             >
               <Text style={[styles.itemText, { color }]}>Open Modal</Text>
             </BlurView>
@@ -135,16 +182,16 @@ export function Settings() {
           visible={isOpenModal}
           onRequestClose={() => setIsOpenModal(false)}
           onDismiss={() => setIsOpenModal(false)}
-          style={StyleSheet.absoluteFill}
+          style={styles.absoluteFill}
         >
           <BlurView
             blurTarget={targetRef}
             type={blurType}
             radius={radius}
-            style={StyleSheet.absoluteFill}
+            style={styles.absoluteFill}
           />
 
-          <View style={[styles.modalContainer, StyleSheet.absoluteFill]}>
+          <View style={[styles.modalContainer, styles.absoluteFill]}>
             <View
               style={[styles.modalContent, isDark && styles.modalContentDark]}
             >
@@ -156,7 +203,7 @@ export function Settings() {
               >
                 On Android platforms, the component utilizes the Dimezis's
                 BlurView library to offer native blur effects with
-                hardware-accelerated rendering.
+                hardware-accelerated rendering ⚡️
               </Text>
 
               <Text
@@ -166,7 +213,7 @@ export function Settings() {
                 ]}
               >
                 On iOS all types are supported by default. However, on Android
-                they are RGBA colors to simulate the same blur color.
+                they are RGBA colors to simulate the same blur color 🎨
               </Text>
 
               <Text
@@ -184,7 +231,7 @@ export function Settings() {
                   isDark && styles.configurationTextDark,
                 ]}
               >
-                Current radius: {radius}.
+                Current radius: {radius} 🌫️
               </Text>
 
               <Text
@@ -193,7 +240,7 @@ export function Settings() {
                   isDark && styles.configurationTextDark,
                 ]}
               >
-                Current blur type: {blurType}.
+                Current blur type: {blurType} 🎨
               </Text>
 
               {isIos && (
@@ -203,7 +250,7 @@ export function Settings() {
                     isDark && styles.configurationTextDark,
                   ]}
                 >
-                  Current effect style: {effectStyle}.
+                  Current effect style: {effectStyle} 💫
                 </Text>
               )}