+++ example/src/screens/Vibrancies/styles.ts
@@ -74,4 +74,13 @@ export const makeStyles = ({ top, bottom }: MakeStylesProps) =>
 
       color: 'white',
     },
+
+    absoluteFill: {
+      position: 'absolute',
+
+      top: 0,
+      left: 0,
+      right: 0,
+      bottom: 0,
+    },
   });