+++ ios/BlurUtils.h
@@ -4,7 +4,6 @@
 
 + (NSNumber *)clipRadius:(NSNumber *)radius;
 + (UIBlurEffectStyle)blurEffectStyle:(NSString *)style;
-+ (UIColor *)colorFromString:(NSString *)colorString;
 
 #if !TARGET_OS_TV
 + (UIVibrancyEffectStyle)vibrancyEffectStyle:(NSString *)style;