+++ ios/BlurView.mm
@@ -4,6 +4,7 @@
 #import "BlurViewEffect.h"
 #import "BlurUtils.h"
 
+#import <react/RCTConversions.h>
 #import <react/renderer/components/BlurViewSpec/ComponentDescriptors.h>
 #import <react/renderer/components/BlurViewSpec/EventEmitters.h>
 #import <react/renderer/components/BlurViewSpec/Props.h>
@@ -46,8 +47,8 @@ using namespace facebook::react;
 
     self.clipsToBounds = YES;
     self.overlayColor = @"light";
-    self.reducedTransparencyFallbackColor = @"white";
-    self.blurRadius = @10;
+    self.reducedTransparencyFallbackColor = [UIColor whiteColor];
+    self.radius = @10;
 
     self.blurEffectView = [[UIVisualEffectView alloc] init];
     self.blurEffectView.autoresizingMask = UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight;
@@ -72,9 +73,9 @@ using namespace facebook::react;
   const auto &oldViewProps = *std::static_pointer_cast<BlurViewProps const>(_props);
   const auto &newViewProps = *std::static_pointer_cast<BlurViewProps const>(props);
 
-  if (oldViewProps.blurRadius != newViewProps.blurRadius) {
-    NSNumber *blurRadius = [NSNumber numberWithDouble:newViewProps.blurRadius];
-    [self setRadius:blurRadius];
+  if (oldViewProps.radius != newViewProps.radius) {
+    NSNumber *radius = [NSNumber numberWithDouble:newViewProps.radius];
+    [self setRadius:radius];
   }
 
   if (oldViewProps.overlayColor != newViewProps.overlayColor) {
@@ -83,7 +84,7 @@ using namespace facebook::react;
   }
 
   if (oldViewProps.reducedTransparencyFallbackColor != newViewProps.reducedTransparencyFallbackColor) {
-    NSString *reducedTransparencyFallbackColor = [NSString stringWithUTF8String:newViewProps.reducedTransparencyFallbackColor.c_str()];
+    UIColor *reducedTransparencyFallbackColor = RCTUIColorFromSharedColor(newViewProps.reducedTransparencyFallbackColor);
     [self setReducedTransparencyFallbackColor:reducedTransparencyFallbackColor];
   }
 
@@ -95,7 +96,7 @@ using namespace facebook::react;
   self.blurEffectView.effect = nil;
 
   UIBlurEffectStyle blurEffectStyle = [BlurUtils blurEffectStyle:self.overlayColor];
-  BlurViewEffect *blurEffect = [BlurViewEffect effectWithStyle:blurEffectStyle andRadius:self.blurRadius];
+  BlurViewEffect *blurEffect = [BlurViewEffect effectWithStyle:blurEffectStyle andRadius:self.radius];
 
   self.blurEffect = blurEffect;
   self.blurEffectView.effect = blurEffect;
@@ -126,7 +127,7 @@ using namespace facebook::react;
     }
   }
 
-  self.reducedTransparencyFallbackView.backgroundColor = [BlurUtils colorFromString:self.reducedTransparencyFallbackColor];
+  self.reducedTransparencyFallbackView.backgroundColor = self.reducedTransparencyFallbackColor;
 }
 
 - (void)reduceTransparencyStatusDidChange:(__unused NSNotification *)notification
@@ -199,13 +200,13 @@ Class<RCTComponentViewProtocol> BlurViewCls(void)
 
 - (void)setRadius:(NSNumber *)radius
 {
-  if (radius && ![self.blurRadius isEqualToNumber:radius]) {
-    _blurRadius = [BlurUtils clipRadius:radius];
+  if (radius && ![self.radius isEqualToNumber:radius]) {
+    _radius = [BlurUtils clipRadius:radius];
     [self updateBlurEffect];
   }
 }
 
-- (void)setReducedTransparencyFallbackColor:(NSString *)reducedTransparencyFallbackColor
+- (void)setReducedTransparencyFallbackColor:(UIColor *)reducedTransparencyFallbackColor
 {
   if (reducedTransparencyFallbackColor && ![self.reducedTransparencyFallbackColor isEqual:reducedTransparencyFallbackColor]) {
     _reducedTransparencyFallbackColor = reducedTransparencyFallbackColor;