+++ src/@types/BlurView.ts
@@ -44,7 +44,7 @@ export interface BlurViewProps extends ViewProps {
 
   /**
    * @description Set the downscale factor for the blur effect. It accepts a
-   * number between `0` and `15`. The higher the value, the more blurred the
+   * number between `0` and `100`. The higher the value, the more blurred the
    * effect will be, but it may also impact performance.
    *
    * @default 6
@@ -65,6 +65,21 @@ export interface BlurViewProps extends ViewProps {
   overlayColor?: ColorValue;
 
   /**
+   * @description Overrides the color type of the overlay on Android. It accepts
+   * a color value that will be used as the background color for the blur effect,
+   * instead of the default color associated with the `type` prop. This can be
+   * useful when you want to achieve a specific look or when you want to match
+   * the blur effect with other elements in your app.
+   *
+   * @default undefined
+   *
+   * @platform Android
+   *
+   * @since 2.2.0
+   */
+  androidColor?: ColorValue;
+
+  /**
    * @description Set a background color about blur effect when reduced
    * transparency is enabled.
    *
@@ -74,5 +89,5 @@ export interface BlurViewProps extends ViewProps {
    *
    * @since 1.4.0
    */
-  reducedTransparencyFallbackColor?: string;
+  reducedTransparencyFallbackColor?: ColorValue;
 }