+++ src/BlurView.tsx
@@ -1,5 +1,12 @@
-import { Children, useMemo, useState, useEffect, useCallback } from 'react';
-import { findNodeHandle, Platform, StyleSheet, View } from 'react-native';
+import {
+  Children,
+  useMemo,
+  useState,
+  useEffect,
+  useCallback,
+  forwardRef,
+} from 'react';
+import { findNodeHandle, Platform, View } from 'react-native';
 
 import Blur from './BlurViewNativeComponent';
 import type { BlurViewProps } from './@types';
@@ -21,7 +28,7 @@ import { globalStyles } from './styles';
  * import { styles } from './styles';
  *
  * const MyComponent = () => {
- *   const blurTargetRef = useRef(null);
+ *   const blurTargetRef = useRef<View | null>(null);
  *
  *   return (
  *     <View style={styles.container}>
@@ -35,7 +42,7 @@ import { globalStyles } from './styles';
  * };
  * ```
  */
-export const BlurView = (props: BlurViewProps) => {
+export const BlurView = forwardRef<View, BlurViewProps>((props, ref) => {
   const {
     type = 'light',
     radius = 10,
@@ -45,6 +52,7 @@ export const BlurView = (props: BlurViewProps) => {
     style,
     children,
     overlayColor,
+    androidColor,
     ...rest
   } = props;
 
@@ -59,32 +67,29 @@ export const BlurView = (props: BlurViewProps) => {
   }, [blurTarget]);
 
   const commonProps = useMemo(() => {
-    /**
-     * When the type is not a primary one, we need to increase the blur radius
-     * and decrease the downscale factor to achieve a similar effect on Android.
-     */
-    const isPrimary =
-      type === 'extra-light' ||
-      type === 'light' ||
-      type === 'dark' ||
-      type === 'extra-dark';
-    const _blurRadius = isPrimary ? radius : 35;
-    const _downscaleFactor = isPrimary
-      ? downscaleFactor
-      : downscaleFactor * 0.66;
+    const isPrimary = 
+        type === 'extra-light' || 
+        type === 'light' || 
+        type === 'dark' || 
+        type === 'extra-dark';
+    const _radius = isPrimary ? radius : 35;
 
     return {
+      ref,
+      androidColor,
+      downscaleFactor,
       targetId: isAndroid ? targetId : undefined,
       reducedTransparencyFallbackColor,
-      downscaleFactor: _downscaleFactor,
       overlayColor: type,
-      blurRadius: isAndroid ? _blurRadius : radius,
+      radius: isAndroid ? _radius : radius,
       ...rest,
     };
   }, [
+    ref,
     type,
     radius,
     downscaleFactor,
+    androidColor,
     reducedTransparencyFallbackColor,
     targetId,
     rest,
@@ -100,7 +105,7 @@ export const BlurView = (props: BlurViewProps) => {
     return (
       <View style={[globalStyles.container, style]}>
         <Blur
-          style={[StyleSheet.absoluteFill, backgroundColor]}
+          style={[globalStyles.absoluteFill, backgroundColor]}
           {...commonProps}
         />
       </View>
@@ -112,18 +117,18 @@ export const BlurView = (props: BlurViewProps) => {
       style={[globalStyles.container, style, !isAndroid && backgroundColor]}
     >
       {isAndroid ? (
-        <Blur style={StyleSheet.absoluteFill} {...commonProps}>
+        <Blur style={globalStyles.absoluteFill} {...commonProps}>
           <View style={[globalStyles.content, style, backgroundColor]}>
             {children}
           </View>
         </Blur>
       ) : (
         <>
-          <Blur style={StyleSheet.absoluteFill} {...commonProps} />
+          <Blur style={globalStyles.absoluteFill} {...commonProps} />
 
-          <View style={globalStyles.content}>{children}</View>
+          {children}
         </>
       )}
     </View>
   );
-};
+});