+++ src/VibrancyViewNativeComponent.ts
@@ -1,4 +1,4 @@
-import type { ViewProps } from 'react-native';
+import type { ColorValue, ViewProps } from 'react-native';
 import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNativeComponent';
 import type {
   Float,
@@ -8,8 +8,8 @@ import type {
 export interface NativeProps extends ViewProps {
   overlayColor?: WithDefault<string, 'light'>;
   effectStyle?: WithDefault<string, 'label'>;
-  blurRadius?: WithDefault<Float, 10.0>;
-  reducedTransparencyFallbackColor?: WithDefault<string, 'white'>;
+  radius?: WithDefault<Float, 10.0>;
+  reducedTransparencyFallbackColor?: ColorValue;
 }
 
 export default codegenNativeComponent<NativeProps>('VibrancyView', {