+++ src/styles/global.ts
@@ -19,4 +19,13 @@ export const globalStyles = StyleSheet.create({
   expand: {
     flex: 1,
   },
+
+  absoluteFill: {
+    position: 'absolute',
+
+    left: 0,
+    right: 0,
+    top: 0,
+    bottom: 0,
+  },
 });